+++ src/maggma/api/resource/read_resource.py
@@ -4,6 +4,8 @@ from typing import Any, Dict, List, Optional, Type, Union
 from fastapi import Depends, HTTPException, Path, Request
 from fastapi import Response
 from pydantic import BaseModel
+from pymongo import timeout as query_timeout
+from pymongo.errors import NetworkTimeout, PyMongoError
 
 from maggma.api.models import Meta
 from maggma.api.models import Response as ResponseModel
@@ -33,6 +35,7 @@ class ReadOnlyResource(Resource):
         key_fields: Optional[List[str]] = None,
         hint_scheme: Optional[HintScheme] = None,
         header_processor: Optional[HeaderProcessor] = None,
+        timeout: Optional[int] = None,
         enable_get_by_key: bool = True,
         enable_default_search: bool = True,
         disable_validation: bool = False,
@@ -47,6 +50,8 @@ class ReadOnlyResource(Resource):
             query_operators: Operators for the query language
             hint_scheme: The hint scheme to use for this resource
             header_processor: The header processor to use for this resource
+            timeout: Time in seconds Pymongo should wait when querying MongoDB
+                before raising a timeout error
             key_fields: List of fields to always project. Default uses SparseFieldsQuery
                 to allow user to define these on-the-fly.
             enable_get_by_key: Enable default key route for endpoint.
@@ -65,6 +70,7 @@ class ReadOnlyResource(Resource):
         self.versioned = False
         self.enable_get_by_key = enable_get_by_key
         self.enable_default_search = enable_default_search
+        self.timeout = timeout
         self.disable_validation = disable_validation
         self.include_in_schema = include_in_schema
         self.sub_path = sub_path
@@ -79,7 +85,10 @@ class ReadOnlyResource(Resource):
             if query_operators is not None
             else [
                 PaginationQuery(),
-                SparseFieldsQuery(model, default_fields=[self.store.key, self.store.last_updated_field],),
+                SparseFieldsQuery(
+                    model,
+                    default_fields=[self.store.key, self.store.last_updated_field],
+                ),
             ]
         )
 
@@ -102,7 +111,9 @@ class ReadOnlyResource(Resource):
         model_name = self.model.__name__
 
         if self.key_fields is None:
-            field_input = SparseFieldsQuery(self.model, [self.store.key, self.store.last_updated_field]).query
+            field_input = SparseFieldsQuery(
+                self.model, [self.store.key, self.store.last_updated_field]
+            ).query
         else:
 
             def field_input():
@@ -110,8 +121,10 @@ class ReadOnlyResource(Resource):
 
         async def get_by_key(
             request: Request,
-            response: Response,
-            key: str = Path(..., alias=key_name, title=f"The {key_name} of the {model_name} to get",),
+            temp_response: Response,
+            key: str = Path(
+                ..., alias=key_name, title=f"The {key_name} of the {model_name} to get",
+            ),
             _fields: STORE_PARAMS = Depends(field_input),
         ):
             f"""
@@ -125,11 +138,29 @@ class ReadOnlyResource(Resource):
             """
             self.store.connect()
 
-            item = [self.store.query_one(criteria={self.store.key: key}, properties=_fields["properties"],)]
+            try:
+                with query_timeout(self.timeout):
+                    item = [
+                        self.store.query_one(
+                            criteria={self.store.key: key},
+                            properties=_fields["properties"],
+                        )
+                    ]
+            except (NetworkTimeout, PyMongoError) as e:
+                if e.timeout:
+                    raise HTTPException(
+                        status_code=504,
+                        detail="Server timed out trying to obtain data. Try again with a smaller request.",
+                    )
+                else:
+                    raise HTTPException(
+                        status_code=500,
+                    )
 
             if item == [None]:
                 raise HTTPException(
-                    status_code=404, detail=f"Item with {self.store.key} = {key} not found",
+                    status_code=404,
+                    detail=f"Item with {self.store.key} = {key} not found",
                 )
 
             for operator in self.query_operators:
@@ -143,7 +174,7 @@ class ReadOnlyResource(Resource):
                 )
 
             if self.header_processor is not None:
-                self.header_processor.process_header(response, request)
+                self.header_processor.process_header(temp_response, request)
 
             return response
 
@@ -163,13 +194,17 @@ class ReadOnlyResource(Resource):
 
         async def search(**queries: Dict[str, STORE_PARAMS]) -> Union[Dict, Response]:
             request: Request = queries.pop("request")  # type: ignore
-            response: Response = queries.pop("temp_response")  # type: ignore
+            temp_response: Response = queries.pop("temp_response")  # type: ignore
 
             query_params = [
-                entry for _, i in enumerate(self.query_operators) for entry in signature(i.query).parameters
+                entry
+                for _, i in enumerate(self.query_operators)
+                for entry in signature(i.query).parameters
             ]
 
-            overlap = [key for key in request.query_params.keys() if key not in query_params]
+            overlap = [
+                key for key in request.query_params.keys() if key not in query_params
+            ]
             if any(overlap):
                 if "limit" in overlap or "skip" in overlap:
                     raise HTTPException(
@@ -193,10 +228,28 @@ class ReadOnlyResource(Resource):
                 query.update(hints)
 
             self.store.connect()
+            try:
+                with query_timeout(self.timeout):
+                    count = self.store.count(
+                        **{
+                            field: query[field]
+                            for field in query
+                            if field in ["criteria", "hint"]
+                        }
+                    )
 
-            count = self.store.count(**{field: query[field] for field in query if field in ["criteria", "hint"]})
+                    data = list(self.store.query(**query))
+            except (NetworkTimeout, PyMongoError) as e:
+                if e.timeout:
+                    raise HTTPException(
+                        status_code=504,
+                        detail="Server timed out trying to obtain data. Try again with a smaller request.",
+                    )
+                else:
+                    raise HTTPException(
+                        status_code=500,
+                    )
 
-            data = list(self.store.query(**query))
             operator_meta = {}
 
             for operator in self.query_operators:
@@ -213,7 +266,7 @@ class ReadOnlyResource(Resource):
                 )
 
             if self.header_processor is not None:
-                self.header_processor.process_header(response, request)
+                self.header_processor.process_header(temp_response, request)
 
             return response