+++ docs/CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## [v0.48.0](https://github.com/materialsproject/maggma/tree/v0.48.0) (2022-08-04)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.6...v0.48.0)
+
+**Merged pull requests:**
+
+- Proposal: remove Drone class [\#669](https://github.com/materialsproject/maggma/pull/669) ([rkingsbury](https://github.com/rkingsbury))
+
+## [v0.47.6](https://github.com/materialsproject/maggma/tree/v0.47.6) (2022-08-04)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.5...v0.47.6)
+
+**Merged pull requests:**
+
+- Docs: add mermaid diagram support [\#677](https://github.com/materialsproject/maggma/pull/677) ([rkingsbury](https://github.com/rkingsbury))
+
+## [v0.47.5](https://github.com/materialsproject/maggma/tree/v0.47.5) (2022-07-26)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.4...v0.47.5)
+
+**Merged pull requests:**
+
+- Add pymongo timeout context to queries [\#691](https://github.com/materialsproject/maggma/pull/691) ([munrojm](https://github.com/munrojm))
+
+## [v0.47.4](https://github.com/materialsproject/maggma/tree/v0.47.4) (2022-07-25)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.3...v0.47.4)
+
+**Merged pull requests:**
+
+- Ensure all fields are properly sanitized [\#690](https://github.com/materialsproject/maggma/pull/690) ([munrojm](https://github.com/munrojm))
+
+## [v0.47.3](https://github.com/materialsproject/maggma/tree/v0.47.3) (2022-06-07)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.2...v0.47.3)
+
+**Merged pull requests:**
+
+- Fix header processing [\#679](https://github.com/materialsproject/maggma/pull/679) ([munrojm](https://github.com/munrojm))
+
+## [v0.47.2](https://github.com/materialsproject/maggma/tree/v0.47.2) (2022-05-27)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.1...v0.47.2)
+
+**Merged pull requests:**
+
+- Docs updates: add FileStore and misc edits [\#668](https://github.com/materialsproject/maggma/pull/668) ([rkingsbury](https://github.com/rkingsbury))
+
+## [v0.47.1](https://github.com/materialsproject/maggma/tree/v0.47.1) (2022-05-24)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.47.0...v0.47.1)
+
+**Merged pull requests:**
+
+- Fix gridfs URI store [\#667](https://github.com/materialsproject/maggma/pull/667) ([utf](https://github.com/utf))
+
 ## [v0.47.0](https://github.com/materialsproject/maggma/tree/v0.47.0) (2022-05-23)
 
 [Full Changelog](https://github.com/materialsproject/maggma/compare/v0.46.2...v0.47.0)
@@ -90,15 +146,15 @@
 
 ## [v0.44.0](https://github.com/materialsproject/maggma/tree/v0.44.0) (2022-03-07)
 
-[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.42.0...v0.44.0)
-
-## [v0.42.0](https://github.com/materialsproject/maggma/tree/v0.42.0) (2022-03-07)
-
-[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.43.0...v0.42.0)
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.43.0...v0.44.0)
 
 ## [v0.43.0](https://github.com/materialsproject/maggma/tree/v0.43.0) (2022-03-07)
 
-[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.41.1...v0.43.0)
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.42.0...v0.43.0)
+
+## [v0.42.0](https://github.com/materialsproject/maggma/tree/v0.42.0) (2022-03-07)
+
+[Full Changelog](https://github.com/materialsproject/maggma/compare/v0.41.1...v0.42.0)
 
 **Merged pull requests:**